Condense project, change pkg tool to poetry
The bindings dir has been deleted, with the webservice and xpcom
dirs being moved. 'webservice' was moved inside vboxapi to be installed along
side the vboxapi pkg. The intention is to get it installed as a separate
package, however the funtionality is still busted due to wsdl generation
into py being based on a library that does py2.

'xpcom' has been moved to the project root to be installed as a separate
package. Originally, sys.path was appended with the location of this dir.
This isn't necessary and is what I consider bad practive.

Finally, distutils was the package management tool used originally for
the vboxapi package. This is now replaced to with the poetry package.
Poetry is a modern Python packaging tool, which also follows
PEP 517.